Code is now a SwitchWithRequirements
Changed variables for arrows with a CodeSprite list.
Created CodeSprite class (stores Up and Down SpriteRenderers).
Added variables for minimum and maximum changes for Randomizer.
Changed UpdateLevelSprites to use Requirement and CodeSprite as arguments.
Changed Randomizer to get a number of changes, randomly change some requirements and update the sprites.