Puzzle Eye with Load/Save + PuzzleEyeStarter
Renamed DestroyMe to PuzzleEyeStarter and moved to Scripts/Other.
Change of camera is now make with PuzzleEyeStarter (instead of CCT).
Stored animators of Doors and Eyes in variables on PuzzleEye.
Simplification of GetComponents for Light2D.
Added methods for DoorDown, EyeStart and EyeNone.
Changed start of puzzle from Update to proper method (StartPuzzle).
Created methods for Reset and EndPuzzle.
Use of nameof in Invokes.